rename room session stream and record instances
