+++ packages/core/src/rooms/RoomSessionStream.test.ts
@@ -13,16 +13,19 @@ describe('RoomSessionStream', () => {
       instance = createRoomSessionStreamObject({
         store: configureJestStore(),
         emitter: new EventEmitter<RoomSessionStreamEventsHandlerMapping>(),
+        payload: {
+          // @ts-expect-error
+          stream: {
+            id: 'c22d7223-5a01-49fe-8da0-46bec8e75e32',
+          },
+          room_session_id: 'room-session-id',
+        },
       })
       // @ts-expect-error
       instance.execute = jest.fn()
     })
 
     it('should control an active stream', async () => {
-      // Mock properties
-      instance.id = 'c22d7223-5a01-49fe-8da0-46bec8e75e32'
-      instance.roomSessionId = 'room-session-id'
-
       const baseExecuteParams = {
         method: '',
         params: {
