+++ packages/core/src/rooms/index.ts
@@ -11,7 +11,5 @@ export interface BaseRoomInterface<
 export * from './methods'
 export * from './methodsRT'
 export * from './RoomSessionRecording'
-export * from './RoomSessionRTRecording'
 export * from './RoomSessionPlayback'
 export * from './RoomSessionStream'
-export * from './RoomSessionRTStream'
