+++ packages/realtime-api/src/video/workers/videoRecordingWorker.ts
@@ -3,7 +3,7 @@ import {
   SagaIterator,
   MapToPubSubShape,
   VideoRecordingEvent,
-  RoomSessionRTRecording,
+  RoomSessionRecording,
   Rooms,
 } from '@signalwire/core'
 import { RoomSession } from '../RoomSession'
@@ -24,9 +24,9 @@ export const videoRecordingWorker = function* (
     throw new Error('Missing room session instance for playback')
   }
 
-  let recordingInstance = get<RoomSessionRTRecording>(payload.recording.id)
+  let recordingInstance = get<RoomSessionRecording>(payload.recording.id)
   if (!recordingInstance) {
-    recordingInstance = Rooms.createRoomSessionRTRecordingObject({
+    recordingInstance = Rooms.createRoomSessionRecordingObject({
       // @ts-expect-error
       store: client.store,
       // @ts-expect-error
@@ -36,7 +36,7 @@ export const videoRecordingWorker = function* (
   } else {
     recordingInstance.setPayload(payload)
   }
-  set<RoomSessionRTRecording>(payload.recording.id, recordingInstance)
+  set<RoomSessionRecording>(payload.recording.id, recordingInstance)
 
   switch (type) {
     case 'video.recording.started':
@@ -45,7 +45,7 @@ export const videoRecordingWorker = function* (
       break
     case 'video.recording.ended':
       roomSessionInstance.baseEmitter.emit(type, recordingInstance)
-      remove<RoomSessionRTRecording>(payload.recording.id)
+      remove<RoomSessionRecording>(payload.recording.id)
       break
     default:
       break
